Fix the variable substitution for task results inside array parameters
(cherry picked from commit 597f71ce4e94c820bde131104e4142c7f9bf85bc)
